# apps/starter-kit/api/swagger.yaml
openapi: 3.0.3
info:
  title: Hackathon - Starter Kit
  description: >
    This Hackathon starter kit is designed to give teams a quick head start 
    on building integrations with Open Finance APIs.
  version: 0.1.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...
paths:
...
  /consent-create/bank-data:
    post:
      summary: Create a bank data authorization request
      description: >
        Generates a consent request for bank data access with specified permissions, 
        signs the request using JWT, and returns a redirect link for authorization.
      tags:
        - Consent Creation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - data_permissions
              properties:
                data_permissions:
                  type: array
                  items:
                    type: string
                    enum:
                      - ReadAccountsBasic
                      - ReadAccountsDetail
                      - ReadBalances
                      - ReadBeneficiariesBasic
                      - ReadBeneficiariesDetail
                      - ReadTransactionsBasic
                      - ReadTransactionsDetail
                      - ReadTransactionsCredits
                      - ReadTransactionsDebits
                      - ReadScheduledPaymentsBasic
                      - ReadScheduledPaymentsDetail
                      - ReadDirectDebits
                      - ReadStandingOrdersBasic
                      - ReadStandingOrdersDetail
                      - ReadConsents
                      - ReadPartyUser
                      - ReadPartyUserIdentity
                      - ReadParty
              example:
                data_permissions:
                  - ReadAccountsBasic
                  - ReadBalances
      responses:
        '200':
          description: Authorization request created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  redirect:
                    type: string
                  consent_id:
                    type: string
                  code_verifier:
                    type: string
        '400':
          description: Invalid request due to missing or incorrect data_permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  description:
                    type: string
                    example: Invalid data_permissions
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"
  /consent-create/single-payment:
    post:
      summary: Create a single instant payment consent
      description: >
        Generates a consent request for a single instant payment of a specified amount,
        signs the request using JWT, and returns a redirect link for authorization.
      tags:
        - Consent Creation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - payment_amount
              properties:
                payment_amount:
                  type: string
                  pattern: "^(?:0|[1-9]\\d*)(\\.\\d{2})$"
                  description: Payment amount in AED with 2 decimal places
              example:
                payment_amount: "100.00"
      responses:
        '200':
          description: Payment consent request created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  redirect:
                    type: string
                  consent_id:
                    type: string
                  code_verifier:
                    type: string
        '400':
          description: Invalid request due to missing or incorrect payment_amount
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid payment_amount"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"
  /consent-create/variable-on-demand-payments:
    post:
      summary: Create a variable on-demand payment consent
      description: >
        Generates a consent request for a variable on-demand payment up to a specified maximum amount,
        signs the request using JWT, and returns a redirect link for authorization.
      tags:
        - Consent Creation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - max_payment_amount
              properties:
                max_payment_amount:
                  type: string
                  pattern: "^(?:0|[1-9]\\d*)(\\.\\d{2})$"
                  description: Maximum payment amount in AED with 2 decimal places
              example:
                max_payment_amount: "100.00"
      responses:
        '200':
          description: Payment consent request created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  redirect:
                    type: string
                  consent_id:
                    type: string
                  code_verifier:
                    type: string
        '400':
          description: Invalid request due to missing or incorrect max_payment_amount
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Invalid max_payment_amount"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"
...